+++ javaProject/transactions/Transaction.java
@@ -4,5 +4,5 @@ public abstract class Transaction {
     public String transactionName;
     public String transactionType;
 
-    abstract public double getAmount(int days);
+    abstract public double getAmount(double days);
 }
